requirements.txt
pandas>=2.0.0
numpy>=1.24.0
apache-iotdb==1.3.3
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
pyod>=1.1.0
statsmodels>=0.14.0
